Data.Conduit.Filesystem

Documentation

traverse

Arguments

:: MonadIO m

=> Bool
Follow directory symlinks (only used on POSIX platforms)

-> FilePath
Root directory

-> Producer m FilePath

Starting at some root directory, traverse the filesystem and enumerate every file (or symlink to a file) found.

Note: the option of whether to follow symlinks is currently only checked on POSIX platforms, as the Win32 package does not support querying symlink status. On Windows, symlinks will always be followed.

sourceFile :: MonadResource m => FilePath -> Producer m ByteString

Same as sourceFile, but uses system-filepath's FilePath type.

sinkFile :: MonadResource m => FilePath -> Consumer ByteString m ()

Same as sinkFile, but uses system-filepath's FilePath type.